fix: restore previewing feature
The previewing feature was broken due to switching CMS to Strapi, which handles static pages and posts differently. This fix enables the user to preview any content by visiting the /preview page and specifying the type (page or post) and the id of the content in the query parameters. For example, /preview?type=page&id=[about-page-id] will preview the about page.